chore: reorganize lib/server and store structure
