+++ src/components/main/results/DetailsDrawer.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState, type JSX } from 'react';
 import useUrlParams from '@/hooks/useUrlParams';
 import CardDetails from './CardDetails';
 import Spinner from '../../shared/ui/Spinner';
-import { useGetAnimeByIdQuery, useRefreshAnimeByIdMutation } from '@/services/apiSlice';
+import { useGetAnimeByIdQuery, useRefreshAnimeByIdMutation } from '@/store/api/apiSlice';
 import { extractErrorMessage } from '@/utils/utils';
 
 const DetailsDrawer = (props: { id: number }): JSX.Element => {
