+++ src/store/slices/selectedSlice.ts
